Moves merge book data code to separate file, distinguish edition and work
